% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\docType{class}
\name{AthenaDriver}
\alias{AthenaDriver}
\alias{AthenaDriver-class}
\alias{show,AthenaDriver-method}
\title{Athena Driver Methods}
\usage{
\S4method{show}{AthenaDriver}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Implementations of pure virtual functions defined in the `DBI` package
for AthenaDriver objects.
}
\keyword{internal}
