% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View.R
\name{AthenaPreviewObject}
\alias{AthenaPreviewObject}
\title{Preview the data in an object.}
\usage{
AthenaPreviewObject(connection, rowLimit, ...)
}
\arguments{
\item{connection}{A connection object, as returned by `dbConnect()`.}

\item{rowLimit}{The maximum number of rows to display.}

\item{...}{Parameters specifying the object.}
}
\value{
A data frame containing the data in the object.
}
\description{
Return the data inside an object as a data frame.
}
\details{
The object to previewed must be specified as one of the arguments
(e.g. \code{table = "employees"}); depending on the driver and underlying
data store, additional specification arguments may be required.
}
\keyword{internal}
