% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integration.R
\name{db_desc}
\alias{db_desc}
\alias{db_desc.AthenaConnection}
\title{S3 implementation of \code{db_desc} for Athena (api version 1).}
\usage{
db_desc.AthenaConnection(x)
}
\arguments{
\item{x}{A \link[=dbConnect]{dbConnect()} object, as returned by \code{dbConnect()}}
}
\value{
Character variable containing Meta Data about query sent to Athena. The Meta Data is returned in the following format:

\code{"Athena <paws version> [<profile_name>@region/database]"}
}
\description{
This is a backend function for dplyr to retrieve meta data about Athena queries. Users won't be required to access and run this function.
}
