% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.athena.R
\docType{package}
\name{paws.athena-package}
\alias{paws.athena}
\alias{paws.athena-package}
\title{paws.athena: Connect to 'AWS Athena' using 'paws' ('DBI' Interface)}
\description{
Designed to be compatible with the R package 'DBI' (Database Interface)
    when connecting to Amazon Web Service ('AWS') Athena <https://aws.amazon.com/athena/>.
    To do this 'R' 'paws' Software Development Kit ('SDK')
    <https://github.com/paws-r/paws> is used as a driver.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DyfanJones/paws.athena}
  \item Report bugs at \url{https://github.com/DyfanJones/paws.athena/issues}
}

}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

}
