% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_translate_env.R
\name{sql_translation}
\alias{sql_translation}
\alias{sql_translation.AthenaConnection}
\title{AWS Athena backend dbplyr SQL translation}
\usage{
sql_translation.AthenaConnection(con)
}
\arguments{
\item{con}{An \code{\linkS4class{AthenaConnection}} object, produced by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\description{
Create s3 implementation of \code{sql_translation} for AWS Athena sql translate environment based off
\href{https://docs.aws.amazon.com/athena/latest/ug/data-types.html}{Athena Data Types} and
\href{https://docs.aws.amazon.com/athena/latest/ug/functions-operators-reference-section.html}{DML Queries, Functions, and Operators}
}
