% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\name{athena}
\alias{athena}
\title{Athena Driver}
\usage{
athena()
}
\value{
\code{athena()} returns a s4 class. This class is used active Athena method for \code{\link[DBI]{dbConnect}}
}
\description{
Driver for an Athena paws connection.
}
\seealso{
\code{\link{dbConnect}}
}
