% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\docType{methods}
\name{dbConnect,AthenaDriver-method}
\alias{dbConnect,AthenaDriver-method}
\alias{dbConnect}
\title{Connect to Athena using python's sdk boto3}
\usage{
\S4method{dbConnect}{AthenaDriver}(drv, aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  schema_name = "default", work_group = NULL, poll_interval = NULL,
  encryption_option = c("NULL", "SSE_S3", "SSE_KMS", "CSE_KMS"),
  kms_key = NULL, profile_name = NULL, role_arn = NULL,
  role_session_name = sprintf("paws-athena-session-\%s",
  as.integer(Sys.time())), duration_seconds = 3600L,
  s3_staging_dir = NULL, region_name = NULL, botocore_session = NULL,
  ...)
}
\value{
\code{dbConnect()} returns a s4 class. This object is used to communicate with AWS Athena.
}
\description{
Connect to Athena using python's sdk boto3
}
\seealso{
\code{\link[DBI]{dbConnect}}
}
