% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integration.R
\name{dbplyr_edition}
\alias{dbplyr_edition}
\alias{dbplyr_edition.AthenaConnection}
\title{Declare which version of dbplyr API is being called.}
\usage{
dbplyr_edition.AthenaConnection(con)
}
\arguments{
\item{con}{A \link{dbConnect} object, as returned by \code{dbConnect()}}
}
\value{
Integer for which version of \code{dbplyr} is going to be used.
}
\description{
Declare which version of dbplyr API is being called.
}
