% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noctua.R
\docType{package}
\name{noctua-package}
\alias{noctua}
\alias{noctua-package}
\title{noctua: Connect to 'AWS Athena' using R 'AWS SDK' 'paws' ('DBI' Interface)}
\description{
Designed to be compatible with the R package 'DBI' (Database Interface)
    when connecting to Amazon Web Service ('AWS') Athena <https://aws.amazon.com/athena/>.
    To do the 'R' AWS Software Development Kit ('SDK') 'paws' 
    <https://github.com/paws-r/paws> is used as a driver.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DyfanJones/noctua}
  \item Report bugs at \url{https://github.com/DyfanJones/noctua/issues}
}

}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

}
