% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options.R
\name{noctua_options}
\alias{noctua_options}
\title{A method to change noctua backend file parser.}
\usage{
noctua_options(file_parser = c("data.table", "vroom"))
}
\arguments{
\item{file_parser}{Method to read and write tables to Athena, currently defaults to data.table}
}
\value{
\code{noctua_options()} returns \code{TRUE}, invisibly.
}
\description{
A method to change noctua backend file parser.
}
\examples{
library(noctua)

# change file parser from default data.table to vroom
noctua_options("vroom")
}
