% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_translate_env.R
\name{sql_translate_env}
\alias{sql_translate_env}
\alias{sql_translate_env.AthenaConnection}
\title{AWS Athena backend dbplyr}
\usage{
sql_translate_env.AthenaConnection(con)
}
\arguments{
\item{con}{An \code{\linkS4class{AthenaConnection}} object, produced by
[DBI::dbConnect()]}
}
\description{
Create s3 implementation of \code{sql_translate_env} for AWS Athena sql translate envirnoment based off
\href{https://docs.aws.amazon.com/athena/latest/ug/data-types.html}{Athena Data Types} and 
\href{https://docs.aws.amazon.com/athena/latest/ug/functions-operators-reference-section.html}{DML Queries, Functions, and Operators}
}
